requirements.txt
Flask
Pillow
numpy
opencv-python-headless
tensorflow
matplotlib
uuid